requirements.txt
flask
flask-cors
gunicorn
feedparser
requests
beautifulsoup4
google-cloud-firestore
google-auth
google-auth-oauthlib
